+++ R/stoxFunctionAttributes.R
@@ -229,7 +229,7 @@ processPropertyFormats <- list(
         ), 
         possibleValues = function(projectPath, BootstrapMethodTable, OutputProcesses) {
             # Get the processes to run:
-            processesSansProcessData <- getProcessesSansProcessData(projectPath, modelName = "baseline", startProcess = BootstrapMethodTable$ProcessName, endProcess = OutputProcesses, return.processIndex = TRUE)
+            processesSansProcessData <- getProcessesSansProcessData(projectPath, modelName = "baseline", startProcess = BootstrapMethodTable$ProcessName, endProcess = OutputProcesses, return.processIndex = TRUE, only.valid = TRUE)
             # Scan through the baseline processes to be run and look for processes with the parameter Seed:
             hasSeed <- sapply(processesSansProcessData$functionParameters, function(x) "Seed" %in% names(x))